Refactor rebootAfterLogin component method to redux action
