+++ client/blocks/login/index.jsx
@@ -15,7 +15,6 @@ import classNames from 'classnames';
 import Gridicon from 'components/gridicon';
 import config from 'config';
 import {
-	getRedirectToSanitized,
 	getRequestNotice,
 	getTwoFactorNotificationSent,
 	isTwoFactorEnabled,
@@ -28,10 +27,9 @@ import { wasManualRenewalImmediateLoginAttempted } from 'state/immediate-login/s
 import { getCurrentOAuth2Client } from 'state/ui/oauth2-clients/selectors';
 import getCurrentQueryArguments from 'state/selectors/get-current-query-arguments';
 import getPartnerSlugFromQuery from 'state/selectors/get-partner-slug-from-query';
-import { recordTracksEventWithClientId as recordTracksEvent } from 'state/analytics/actions';
+import { rebootAfterLogin } from 'state/login/actions';
 import { isCrowdsignalOAuth2Client, isWooOAuth2Client } from 'lib/oauth2-clients';
 import { login } from 'lib/paths';
-import userFactory from 'lib/user';
 import Notice from 'components/notice';
 import AsyncLoad from 'components/async-load';
 import VisitSite from 'blocks/visit-site';
@@ -46,8 +44,6 @@ import { isWebAuthnSupported } from 'lib/webauthn';
  */
 import './style.scss';
 
-const user = userFactory();
-
 class Login extends Component {
 	static propTypes = {
 		disableAutoFocus: PropTypes.bool,
@@ -59,8 +55,6 @@ class Login extends Component {
 		linkingSocialService: PropTypes.string,
 		oauth2Client: PropTypes.object,
 		privateSite: PropTypes.bool,
-		recordTracksEvent: PropTypes.func.isRequired,
-		redirectTo: PropTypes.string,
 		requestNotice: PropTypes.object,
 		socialConnect: PropTypes.bool,
 		socialService: PropTypes.string,
@@ -177,23 +171,10 @@ class Login extends Component {
 		this.setState( { continueAsAnotherUser: true } );
 	};
 
-	rebootAfterLogin = async () => {
-		this.props.recordTracksEvent( 'calypso_login_success', {
-			two_factor_enabled: this.props.twoFactorEnabled,
+	rebootAfterLogin = () => {
+		this.props.rebootAfterLogin( {
 			social_service_connected: this.props.socialConnect,
 		} );
-
-		// Redirects to / if no redirect url is available
-		const url = this.props.redirectTo || '/';
-
-		// User data is persisted in localstorage at `lib/user/user` line 157.
-		// Only clear the data if a user is currently set, otherwise keep the
-		// logged out state around so that it can be used in signup.
-		if ( user.get() ) {
-			await user.clear();
-		}
-
-		window.location.href = url;
 	};
 
 	renderHeader() {
@@ -466,7 +447,6 @@ class Login extends Component {
 export default connect(
 	( state ) => ( {
 		currentUser: getCurrentUser( state ),
-		redirectTo: getRedirectToSanitized( state ),
 		requestNotice: getRequestNotice( state ),
 		twoFactorEnabled: isTwoFactorEnabled( state ),
 		twoFactorNotificationSent: getTwoFactorNotificationSent( state ),
@@ -480,5 +460,5 @@ export default connect(
 			'woocommerce-onboarding' === get( getCurrentQueryArguments( state ), 'from' ),
 		wccomFrom: get( getCurrentQueryArguments( state ), 'wccom-from' ),
 	} ),
-	{ recordTracksEvent }
+	{ rebootAfterLogin }
 )( localize( Login ) );
